Files associated with OTOC well organized
